feat: Refactor booking, pricing, and entity mappings
Refactor the booking process, updating DTOs and mappers for improved clarity and maintainability. Introduce new `PublicEndpointConfig` and `PricingDataInitializer` for better API and pricing management. Update security configuration to use the new public endpoint config and adjust access rules. Improve showtime logic to validate for overlapping movie durations using `TIMESTAMPADD(MINUTE,...)` instead of `HOUR, CAST(...)`. Add new fields to Showtime entity to correctly map seat locks and bookings. Update test cases in `BookingIntegrationTest.java` to reflect changes in `LockSeatsRequest` and remove `@PreAuthorize` from controller tests. Remove the old `PricingDataInitializer.java` file as it's now replaced by the one under `configs`. Also update `.gitignore` for new testing directory.

+++ backend/src/main/java/com/api/moviebooking/services/ShowtimeService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.api.moviebooking.helpers.exceptions.ResourceNotFoundException;
+import com.api.moviebooking.helpers.exceptions.EntityDeletionForbiddenException;
 import com.api.moviebooking.helpers.mapstructs.ShowtimeMapper;
 import com.api.moviebooking.models.dtos.showtime.AddShowtimeRequest;
 import com.api.moviebooking.models.dtos.showtime.ShowtimeDataResponse;
@@ -64,7 +65,6 @@ public class ShowtimeService {
         return movieRepo.findById(movieId)
                 .orElseThrow(() -> new ResourceNotFoundException("Movie", "id", movieId));
     }
-
 
     /**
      * validateNoOverlap:
@@ -164,6 +164,19 @@ public class ShowtimeService {
     @Transactional
     public void deleteShowtime(UUID showtimeId) {
         Showtime showtime = findShowtimeById(showtimeId);
+
+        if (!showtime.getSeatLocks().isEmpty()) {
+            throw new EntityDeletionForbiddenException();
+        }
+
+        if (!showtime.getShowtimeSeats().isEmpty()) {
+            throw new EntityDeletionForbiddenException();
+        }
+
+        if (!showtime.getBookings().isEmpty()) {
+            throw new EntityDeletionForbiddenException();
+        }
+
         showtimeRepo.delete(showtime);
     }
